content/base/test/chrome/test_document_register.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=783129
-->
<window title="Mozilla Bug 549682"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=783129"
     target="_blank">Mozilla Bug 783129</a>
  <iframe onload="startTests()" id="fooframe" src="http://example.com"></iframe>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

  /** Test for Bug 783129 **/
  SimpleTest.waitForExplicitFinish();

  function startTests() {
    var c = $("fooframe").contentDocument.register("x-foo");
    var elem = new c();
    is(elem.tagName, "X-FOO", "Constructor should create an x-foo element.");

    var anotherElem = $("fooframe").contentDocument.createElement("x-foo");
    is(anotherElem.tagName, "X-FOO", "createElement should create an x-foo element.");
    SimpleTest.finish();
  }

  ]]></script>
</window>
